requirements.txt
Flask
Flask-SQLAlchemy
openai
python-dotenv
